+++ packages/create/CHANGELOG.md
@@ -1,5 +1,11 @@
 # create-dovenv
 
+## 0.0.19
+
+### Patch Changes
+
+- Docs: fix build issues
+
 ## 0.0.16
 
 ### Patch Changes