+++ packages/plugin/docs/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @dovenv/docs
 
+## 0.0.19
+
+### Patch Changes
+
+- Docs: fix build issues
+
+- Updated dependencies []:
+  - @dovenv/core@0.0.19
+
+## 0.0.18
+
+### Patch Changes
+
+- Docs: Fix build process in ../ paths
+
+- Updated dependencies []:
+  - @dovenv/core@0.0.18
+
 ## 0.0.17
 
 ### Patch Changes