+++ packages/plugin/workspace/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @dovenv/workspace
 
+## 0.0.19
+
+### Patch Changes
+
+- Docs: fix build issues
+
+- Updated dependencies []:
+  - @dovenv/core@0.0.19
+
+## 0.0.18
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @dovenv/core@0.0.18
+
 ## 0.0.17
 
 ### Patch Changes